add automatique update in app import
sh: line 1: q: command not found
