+++ testing/fake-llm-server/githubHandler.ts
@@ -26,6 +26,10 @@ const mockUser = {
   email: "testuser@example.com",
 };
 
+const mockReposRoot = fs.mkdtempSync(
+  path.join(os.tmpdir(), "dyad-git-mock-"),
+);
+
 const mockRepos = [
   {
     id: 1,
@@ -354,31 +358,20 @@ export function handleClearPushEvents(req: Request, res: Response) {
 export function handleGitPush(req: Request, res: Response, next?: Function) {
   console.log("* GitHub Git operation requested:", req.method, req.url);
 
-  // Log request headers to see git operation details
   console.log("* Git Headers:", {
     "git-protocol": req.headers["git-protocol"],
     "content-type": req.headers["content-type"],
     "user-agent": req.headers["user-agent"],
   });
 
-  // Create a unique temporary directory for this request
-  const mockReposRoot = fs.mkdtempSync(
-    path.join(
-      os.tmpdir(),
-      "dyad-git-mock-" + Math.random().toString(36).substring(2, 15),
-    ),
-  );
-  console.error(`* Created temporary git repos directory: ${mockReposRoot}`);
+  console.error(`* Using git repos directory: ${mockReposRoot}`);
 
-  // Create git middleware instance for this request
   const gitHttpMiddleware = gitHttpMiddlewareFactory({
     root: mockReposRoot,
     route: "/github/git",
     glob: "*.git",
   });
 
-  // Extract repo name from URL path like /github/git/testuser/test-repo.git
-  // The middleware expects the repo name as the basename after the route
   const urlPath = req.url;
   const match = urlPath.match(/\/github\/git\/[^/]+\/([^/.]+)\.git/);
   const repoName = match?.[1];
@@ -386,24 +379,18 @@ export function handleGitPush(req: Request, res: Response, next?: Function) {
   if (repoName) {
     console.log(`* Git operation for repo: ${repoName}`);
 
-    // Track push events if this is a git-receive-pack (push) operation
     if (req.url.includes("/git-receive-pack") && req.method === "POST") {
       console.log("* Git PUSH operation detected for repo:", repoName);
 
-      // Collect request body to parse git protocol
       let body = "";
       req.on("data", (chunk) => {
         body += chunk.toString();
       });
       req.on("end", () => {
         try {
-          // Parse git pack protocol for branch refs
-          // Git protocol sends refs in format: "old-sha new-sha refs/heads/branch-name"
           const lines = body.split("\n");
           lines.forEach((line) => {
-            // Look for lines containing refs/heads/
             const refMatch = line.match(
-              // eslint-disable-next-line
               /([0-9a-f]{40})\s+([0-9a-f]{40})\s+refs\/heads\/([^\s\u0000]+)/,
             );
             if (refMatch) {
@@ -434,27 +421,58 @@ export function handleGitPush(req: Request, res: Response, next?: Function) {
       });
     }
 
-    // Ensure the bare git repository exists for this repo
     const bareRepoPath = path.join(mockReposRoot, `${repoName}.git`);
-    console.log(`* Creating bare git repository at: ${bareRepoPath}`);
-    try {
-      fs.mkdirSync(bareRepoPath, { recursive: true });
-      // Initialize as bare repository
-      const { execSync } = require("child_process");
-      execSync(`git init --bare`, { cwd: bareRepoPath });
-      console.log(
-        `* Successfully created bare git repository: ${repoName}.git`,
-      );
-    } catch (error) {
-      console.error(`* Failed to create bare git repository:`, error);
-      return res.status(500).json({
-        message: "Failed to initialize git repository",
-        error: error instanceof Error ? error.message : String(error),
-      });
+    if (!fs.existsSync(bareRepoPath)) {
+      console.log(`* Creating bare git repository at: ${bareRepoPath}`);
+      try {
+        fs.mkdirSync(bareRepoPath, { recursive: true });
+        const { execSync } = require("child_process");
+        execSync(`git init --bare`, { cwd: bareRepoPath });
+
+        const tmpClone = fs.mkdtempSync(
+          path.join(os.tmpdir(), "dyad-git-clone-"),
+        );
+        try {
+          execSync(`git clone "${bareRepoPath}" "${tmpClone}"`, {
+            stdio: "pipe",
+          });
+          fs.writeFileSync(path.join(tmpClone, "README.md"), `# ${repoName}\n`);
+          execSync(`git add README.md`, { cwd: tmpClone, stdio: "pipe" });
+          execSync(
+            `git -c user.name=dyad -c user.email=dyad@example.com commit -m "initial commit"`,
+            { cwd: tmpClone, stdio: "pipe" },
+          );
+          execSync(`git push origin HEAD:refs/heads/main`, {
+            cwd: tmpClone,
+            stdio: "pipe",
+          });
+          try {
+            execSync(
+              `git --git-dir="${bareRepoPath}" symbolic-ref HEAD refs/heads/main`,
+              { stdio: "pipe" },
+            );
+          } catch (err) {
+            console.warn(
+              "* Warning: failed to set symbolic-ref HEAD on bare repo",
+              err,
+            );
+          }
+        } finally {
+          fs.rmSync(tmpClone, { recursive: true, force: true });
+        }
+
+        console.log(
+          `* Successfully created bare git repository: ${repoName}.git`,
+        );
+      } catch (error) {
+        console.error(`* Failed to create bare git repository:`, error);
+        return res.status(500).json({
+          message: "Failed to initialize git repository",
+          error: error instanceof Error ? error.message : String(error),
+        });
+      }
     }
 
-    // Rewrite the URL to match what the middleware expects
-    // Change /github/git/testuser/test-repo.git/... to /github/git/test-repo.git/...
     const rewrittenUrl = req.url.replace(
       /\/github\/git\/[^/]+\//,
       "/github/git/",
@@ -463,13 +481,11 @@ export function handleGitPush(req: Request, res: Response, next?: Function) {
     console.log(`* Rewritten URL from ${urlPath} to ${rewrittenUrl}`);
   }
 
-  // Use git-http-mock-server middleware to handle the actual git operations
   gitHttpMiddleware(
     req,
     res,
     next ||
       (() => {
-        // Fallback if middleware doesn't handle the request
         console.log(
           `* Git middleware did not handle request: ${req.method} ${req.url}`,
         );
